calendar view permissions
Added a decorator @check_calendar_permissions to the calendar views.
This decorator check the new variable CALENDAR_VIEW_PERM in settings
which is either True or False to to decide if the
CHECK_CALENDAR_PERM_FUNCTION also effects the views of the calendars and
not just the event editing on those calendars.