fixed issue #123 and issue #137
the issue was that when events are modified, the persisted occurrences
weren’t having their “original_start” and “original_end” being updated,
so the occurrence replacer wasn’t catching them.

To fix this, I calculated the change of time to the start and end of
the event, and add this to the occurrences, so they maintain their
order in the list of occurrences.