Simplify interface to CalendarByPeriodsView to accept a single period.
All uses of CalendarByPeriodsView a single period is passed. However,
the interface requires a list of periods. As this is always used as a
scalar, might as well pass it as a scalar. This also simplifies
templates as they can now use the period object as a scalar instead of
indexing a dictionary.