Add tox.ini to allow testing the full test matrix locally
tox is a virtualenv management and test command line tool. It is a
powerful and commonly used tool for running tests across multiple
environments. This allows locally running the project's tests against
the full test matrix of supported Python and Django versions with a
single command.

For additional information on using tox, see:

https://tox.readthedocs.io/